suppress warning on 32-bit builds on OS X
git-svn-id: https://svn.r-project.org/R/trunk@59113 00db46b3-68df-0310-9c12-caf00c1e9a41

+++ src/main/qsort-body.c
@@ -50,7 +50,7 @@ L10:
     L20:
         k = i;
         /* ij = (j + i) >> 1; midpoint */
-        ij = i + (INT)((j - i) * R);
+        ij = (INT)(i + (INT)((j - i) * R));
 #ifdef qsort_Index
         it = I[ij];
 #endif
